pyproject.toml
[project]
name = "graph-blas-triag-counting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "networkx>=3.6.1",
    "python-graphblas>=2025.2.0",
    "scipy>=1.17.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.7",
]

[tool.pytest.ini_options]
pythonpath = ["."]
